[visitor] Split TTVisitor into its own module
